requirements.txt
psycopg2
python-dotenv
openai
requests
streamlit
plotly
pandas
bse
feedparser
pyyaml
